Added log.cpp, use new exception macro, cleaned up copyright notices.
Moved log.cpp from advgetopt to cppthread so that way we do not have to
depend on advgetopt which in turn can depend on cppthread as it needs to
lock a few things for our multithread support (which was close to
non-existant in our previous version). The multithread support is very
important for configuration files and thus the getopt class requires it
too. Instead of reinventing the wheel, having access to cppthread gives
us full access to our advance lock classes in cppthread.
Removed the word "exception" from most exceptions, "error" once is
generally enough.
Fixed some references to the "snap" namespace, here we use "cppthread".
Updated the throws and logs as required.
Added a definition for an undefined pid_t (`PID_UNDEFINED`).
Added a definition for an undefined thread_t (`THREAD_UNDEFINED`).
Removed references to "Snap Websites Server".