Added the worker item_with_predicate class.
Changed the FIFO in a deque so we can handle the predicate.
Added an out of range exception for the pool to raise on invalid size/indices.
Fixed a couple of function that used '.' instead of '->'.
Fixed up the copyright notice.
Ran spellcheck against comments.
Fixed the TODO.txt file.