Added support to load plugins at runtime.
Added info about the plugins support to the README.md file.
Fixed the exception documentaton.
Added a few exception for the plugin implementation.
Added a function to load the boot-id from the /proc/... file.
Added a tool to verify the boot-id implementation.
Fixed info about the FIFO in regard to thread pools.