Added a test of the plugins objects.
Added the SNAPDEV_INCLUDE_DIRS to a couple of CMakefiles.txt files.
Fixed a #include that used "..." instead of <...>.
Got a version of the plugins that do get loaded (one single plugin at the moment).
Commented out the plugin_factory destructor error message. Plugins do not get unloaded properly yet.
Added some missing plugin::...() functions.
Moved the "allowed_redirects" to a plugin_paths field instead of just a parameter.
Fixed the segments.erase() code, the index has to be decremented too.
Fixed the plugin_paths::push() implementation, it is expected to canonicalize the input path.
Fixed the to_filename(), the local path needs to have a slash ("./").
Added proper support of the dependencies in the load_plugins() function.
Fixed the name validation (the constexpr version), the size includes the '\0', also the errors have to be thrown, not use the static_assert().
Fixed the out of range error for the date, I used `>` instead of `<`.
Added the library version to the plugin.
Fixed the strings in the plugin_...() initialization functions.
Fixed the CPPTHREAD_PLUGIN_... macros.
Fixed the coverage script so it runs, but there are still issues.
Renamed the tests as "catch_<name>.cpp" so that way it's easier to distinguish from non-catch files.
Added one test plugin. We need more to verify dependencies and conflicts.
Used CATCH_CHECK() instead of CATCH_REQUIRE() in various places.
Added the bootstrap() function to the test plugin.
Various clean ups.