Added a callback to the stop() function.
At times, I want to ask the thread to stop, but I cannot safely do it before
the f_stopping flag is set to true. To help with that, I now have a callback
which gets called after the f_stopping flag is set and before we join() with
the child.