Mask all signals by default.
This should prevent signals from breaking the software as that way it is
much more likely that the signals will be sent to the main process instead
of one of its sub-thread. I have had some issues with a test until I
implemented this, so I think it helps. The signal masking/unmasking is
implemented in the libexcept library (a different project).